==> code/preprocessing/0_raw_data_aggregation.ipynb <==
import pandas as pd
import numpy as np
from utils import *
from sklearn.preprocessing import StandardScaler, MinMaxScaler

import warnings
warnings.filterwarnings('ignore')

years = ['0304', '0506', '0708', '0910', '1112', '1314', '1516', '1718', '1720']
year_char = 'C'

### 1. Collect demographic information

type_demo = 'demographic'
df_demo = concat_data_across_years(type_demo, 'DEMO', years, year_char)
# 95872 unique records in total

# Select the wanted columns. Make changes here if needed in the future.
df_demo = df_demo[['SEQN', 'RIAGENDR', 'RIDAGEYR', 'RIDRETH1', 'INDHHINC', 'DMDEDUC2', 'WTINT2YR',
                       'WTMEC2YR', 'WTINTPRP', 'WTMECPRP', 'years']]
df_demo = df_demo.fillna(0)
df_demo[['SEQN', 'RIAGENDR', 'RIDAGEYR', 'RIDRETH1', 'INDHHINC', 'DMDEDUC2']] = \
    df_demo[['SEQN', 'RIAGENDR', 'RIDAGEYR', 'RIDRETH1', 'INDHHINC', 'DMDEDUC2']].astype(int)

# Unify the weight columns and make column names readable.
df_demo['weight_interview'] = np.where(df_demo['WTINT2YR'] == -1, df_demo['WTINTPRP'], df_demo['WTINT2YR'])
df_demo['weight_mec'] = np.where(df_demo['WTMEC2YR'] == -1, df_demo['WTMECPRP'], df_demo['WTMEC2YR'])
df_demo.drop(['WTINT2YR','WTMEC2YR', 'WTINTPRP', 'WTMECPRP'], axis=1, inplace=True)
df_demo = df_demo.rename(columns={'RIAGENDR': 'gender', 'RIDAGEYR': 'age', 'RIDRETH1': 'race', 'DMDEDUC2': 'education',
                                  'INDHHINC': 'household_income'})
df_demo = df_demo.set_index('SEQN')

# Transform ages to age groups
bins = [-1, 10, 20, 30, 40, 50, 60, 100]
labels = ['1', '2', '3', '4', '5', '6', '7']

# Create a new column for age groups
df_demo['age_group'] = pd.cut(df_demo['age'], bins=bins, labels=labels, right=True)
df_demo.drop('age', axis=1, inplace=True)

### 2. Questionnaire Data

type_questionnaire = 'questionnaire'

#### 2.1. Diet Behaviors & Nutrition

# For simplicity, I use the acronym of each initials as the variable name, so Diet Behaviors & Nutrition is DBN
df_DBN = concat_data_across_years(type_questionnaire, 'DBQ', years, year_char)

df_DBN = df_DBN[['SEQN', 'DBD090', 'DBD091', 'DBD197', 'DBQ197', 'DBQ229', 'DBD895', 'DBQ700',
                 'DBD900', 'DBD905', 'DBD910']]
df_DBN = df_DBN.fillna(-1)
df_DBN = df_DBN.astype(int)

# These 3 columns are the same questions of different field names over the years
df_DBN['meals_outside'] = np.where(df_DBN['DBD090'] == -1, df_DBN['DBD091'], df_DBN['DBD090'])
df_DBN['meals_outside'] = np.where(df_DBN['meals_outside'] == -1, df_DBN['DBD895'], df_DBN['meals_outside'])
# For those who had meals outside more than 21 time a week, we mark it to 22.
# For those who had meals outside less than once a week, we mark it 0.
df_DBN['meals_outside'].loc[df_DBN['meals_outside'] == 5555] = 22
df_DBN['meals_outside'].loc[df_DBN['meals_outside'] == 6666] = 0
df_DBN.drop(['DBD090', 'DBD091', 'DBD895'], axis=1, inplace=True)

df_DBN['milk_consumption_last_30_days'] = np.where(df_DBN['DBD197'] == -1, df_DBN['DBQ197'], df_DBN['DBD197'])
df_DBN.drop(['DBD197', 'DBQ197'], axis=1, inplace=True)

df_DBN = df_DBN.rename(columns={'DBQ229': 'regular_milk_drinker', 'DBQ700': 'self_perceived_diet',
                                'DBD900': 'fastfood_weekly', 'DBD905': 'ready_to_eat_food_monthly',
                                'DBD910': 'frozen_meal_monthly', 'meals_outside': 'meals_outside_weekly'})

# The same code might indicate different things. For weekly data, 6666 is less than once a week.
# For monthly data, 6666 is more than 90 times a month.
df_DBN['fastfood_weekly'].loc[df_DBN['fastfood_weekly'] == 5555] = 22
df_DBN.replace([7777, 9999], -1, inplace=True)
df_DBN.replace([6666], 91, inplace=True)

df_DBN = df_DBN.set_index('SEQN')
df_DBN

df_DBN.to_csv('../processed_data/diet_habits.csv')

#### 2.2. Drug Use

# No specific illicit drug info in 03-04 data. No such table in 17-20 data.
df_DU = concat_data_across_years(type_questionnaire, 'DUQ',
                                 ['0506', '0708', '0910', '1112', '1314', '1516', '1718'], 'D')

df_DU = df_DU.loc[df_DU['DUQ290'] == 1]
df_DU = df_DU[['SEQN', 'DUQ270U', 'DUQ350U', 'DUQ300', 'DUQ310Q', 'DUQ310U']]
df_DU = df_DU.fillna(-1)
df_DU = df_DU.astype(int)

"""
# We only care about those who have used heroin at least once.
# If within a year, the user has been using any illicit drugs (heroin, meth, cocaine),
# we identify the user as an active user.
# Otherwise, we consider this user a recovered user.

# To be specific, -1 means missing, 4 means it has been years that a user haven't used a drug.
# If there is a value that is neither -1, nor 4 in any of the three columns, the user is an active user.

# We label the active user as 1 and recovered user as 2.
# And the rest of the users who hasn't even used heroin or other opioid prescription drugs as 0: non-opioid-user
"""
df_DU['active_user'] = np.where((df_DU['DUQ270U'].isin([4, -1]) == False) |
                                 (df_DU['DUQ350U'].isin([4, -1]) == False) |
                                 (df_DU['DUQ310U'].isin([4, -1]) == False), 1, 2)

df_DU = df_DU.rename(columns={'DUQ300': 'age_first_use_heroin', 'DUQ310U': 'last_time_unit_used_heroin',
                              'DUQ310Q': 'last_time_used_heroin', 'DUQ270U': 'last_time_unit_used_cocaine',
                              'DUQ350U': 'last_time_unit_used_meth'})

"""
last_time_unit_used_heroin and last_time_used_heroin is a combo feature.
time unit defines whether it's years, months or days we are talking about.
the time defines the exact number of that unit.
For example, if time unit is 4 and time is 30, it means this user last used heroin 30 years ago.
"""
df_DU = df_DU.set_index('SEQN')
df_DU

# df_drug_age = df_demo.merge(df_DU, how='right', left_on=df_demo.index, right_on=df_DU.index)

#### 2.3.Prescription Medicine

df_PM = concat_data_across_years(type_questionnaire, 'RXQ_RX', years, year_char)

df_PM = df_PM[df_PM['RXDUSE']==1]

df_PM['SEQN'] = df_PM['SEQN'].astype(int).astype(str)
df_PM[['SEQN', 'RXDDRUG', 'RXDDRGID', 'RXDDAYS']].to_csv('../processed_data/user_prescription_medicine.csv', index=False)


df_PM

df_PM_1 = df_PM.loc[(df_PM['RXDRSC1'] == 'F11.2') | (df_PM['RXDRSC1'] == 'F11.23')]

drugs = pd.read_sas('../data/RXQ_DRUG.xpt', encoding='ISO-8859-1')
drug_60 = drugs[(drugs['RXDDCI1A'] == 57) & (drugs['RXDDCI1B'] == 58) & (drugs['RXDDCI1C'] == 60)]
drug_191 = drugs[(drugs['RXDDCI1A'] == 57) & (drugs['RXDDCI1B'] == 58) & (drugs['RXDDCI1C'] == 191)]
drug = pd.concat([drug_60, drug_191])
drug_id = set(drug['RXDDRGID'].tolist())

df_PM  = df_PM [df_PM ['RXDDRGID'].isin(drug_id)]

df_PM = pd.concat([df_PM, df_PM_1])
df_PM = df_PM.drop_duplicates()

# 3992 records for taking opioid prescription or taking prescription for opioid dependence.
len(df_PM)

df_PM = df_PM[['SEQN', 'RXDDRUG', 'RXDDRGID', 'RXDDAYS']]
df_PM = df_PM[df_PM['RXDDAYS'] > 90]
df_PM[['SEQN', 'RXDDAYS']] = df_PM[['SEQN', 'RXDDAYS']].astype(int)

"""
# We define the long term opioid users as those how have taken opioid prescriptions over 90 days.

# Note that each user can take multiple opioid prescriptions.

# In an earlier study for tracking long term opioid users,
# The author excluded medications containing buprenorphine since they are used to treat use disorder.
# However, we find multiple cases that this medicine used to treat opioid dependence.
# So technically this also implies the user is a long term opioid user.
"""
df_PM = df_PM.rename(columns={'RXDDRUG': 'drug_name', 'RXDDRGID': 'drug_id', 'RXDDAYS': 'days_using'})
df_PM = df_PM.set_index('SEQN')
df_PM

#### Create labels for all users in the main table.

opioid_user_set = set(df_PM.index.tolist())

df_DU['active_user'].value_counts()

"""
Following the labeling scheme, we get the label in the main table.
0: user, 1: active user, 2: recovered_user
"""
df_demo['label'] = 0
df_demo['label'] = df_demo.index.map(df_DU['active_user']).fillna(0)
df_demo.loc[df_demo.index.isin(opioid_user_set), 'label'] = 1

df_demo['label'].value_counts()

df_demo.to_csv('../processed_data/main_table.csv')

### 3. Laboratory Data

type_laboratory = 'laboratory'

#### 3.1. Standard Biochemistry Profile

df_SBP = concat_data_across_years(type_laboratory, 'BIOPRO',
                                 ['0506', '0708', '0910', '1112', '1314', '1516', '1718', '1720'], 'D')
df_temp = concat_data_across_years(type_laboratory, 'L40',
                                 ['0304'], 'C')
df_SBP = pd.concat([df_SBP, df_temp])

# We keep all the items in this laboratory but filter out the same value with different unit.
# For example, we keep "Total calcium (mmol/L)" and get rid of "Total calcium (mg/dL)"
columns_SBP = [
    'SEQN',
    'LBDSALSI', 'LBXSATSI', 'LBXSASSI', 'LBXSAPSI', 'LBDSBUSI', 'LBDSCASI',
    'LBDSCHSI', 'LBXSC3SI', 'LBDSCRSI', 'LBXSGTSI', 'LBDSGLSI', 'LBDSIRSI',
    'LBXSLDSI', 'LBDSPHSI', 'LBDSTBSI', 'LBDSTPSI', 'LBDSTRSI', 'LBDSUASI',
    'LBDSCRSI', 'LBXSNASI', 'LBXSKSI', 'LBXSCLSI', 'LBXSOSSI', 'LBDSGBSI'
]
rename_SBP = {
    'LBDSALSI': 'Albumin (g/L)', 'LBXSATSI': 'Alanine aminotransferase (ALT) (U/L)',
    'LBXSASSI': 'Aspartate aminotransferase (AST) (U/L)', 'LBXSAPSI': 'Alkaline phosphatase (U/L)',
    'LBDSBUSI': 'Blood urea nitrogen (mmol/L)', 'LBDSCASI': 'Total calcium (mmol/L)',
    'LBDSCHSI': 'Cholesterol (mmol/L)', 'LBXSC3SI': 'Bicarbonate (mmol/L)',
    'LBDSCRSI': 'Creatinine (µmol/L)', 'LBXSGTSI': 'Gamma glutamyl transferase (U/L)',
    'LBDSGLSI': 'Glucose, serum (mmol/L)', 'LBDSIRSI': 'Iron, refrigerated (umol/L)',
    'LBXSLDSI': 'Lactate dehydrogenase LDH (U/L)', 'LBDSPHSI': 'Phosphorus (mmol/L)',
    'LBDSTBSI': 'Bilirubin, total (umol/L)', 'LBDSTPSI': 'Total protein (g/L)',
    'LBDSTRSI': 'Triglycerides (mmol/L)', 'LBDSUASI': 'Uric acid (umol/L)',
    'LBXSNASI': 'Sodium (mmol/L)',
    'LBXSKSI': 'Potassium (mmol/L)', 'LBXSCLSI': 'Chloride (mmol/L)',
    'LBXSOSSI': 'Osmolality (mmol/Kg)', 'LBDSGBSI': 'Globulin (g/L)'
}

# Not every respondent take this examination. In the 95872 respondents from 2003-2020, we have 60331 valid records.
df_SBP = df_SBP[columns_SBP]
df_SBP = df_SBP.rename(columns=rename_SBP)
df_SBP['SEQN'] = df_SBP['SEQN'].astype(int)
df_SBP = df_SBP.set_index('SEQN')
df_SBP.dropna(how='all', inplace=True)
df_SBP

# 89% of the active users have this data; 93% of the recovered users have this data.
check_coverage(df_demo, df_SBP)

#### 3.2 Complete Blood Count

df_CBC = concat_data_across_years(type_laboratory, 'CBC',
                                 ['0506', '0708', '0910', '1112', '1314', '1516', '1718', '1720'], 'D')
df_temp = concat_data_across_years(type_laboratory, 'L25',
                                 ['0304'], 'C')
df_CBC = pd.concat([df_CBC, df_temp])

columns_CBC = [
    'SEQN',
    'LBXWBCSI', 'LBXLYPCT', 'LBXMOPCT', 'LBXNEPCT', 'LBXEOPCT',
    'LBXBAPCT', 'LBDLYMNO', 'LBDMONO', 'LBDNENO', 'LBDEONO',
    'LBDBANO', 'LBXRBCSI', 'LBXHGB', 'LBXHCT', 'LBXMCVSI',
    'LBXMCHSI', 'LBXMC', 'LBXRDW', 'LBXPLTSI', 'LBXMPSI'
]
rename_CBC = {
    'LBXWBCSI': 'White blood cell count (1000 cells/uL)',
    'LBXLYPCT': 'Lymphocyte percent (%)',
    'LBXMOPCT': 'Monocyte percent (%)',
    'LBXNEPCT': 'Segmented neutrophils percent (%)',
    'LBXEOPCT': 'Eosinophils percent (%)',
    'LBXBAPCT': 'Basophils percent (%)',
    'LBDLYMNO': 'Lymphocyte number (1000 cells/uL)',
    'LBDMONO': 'Monocyte number (1000 cells/uL)',
    'LBDNENO': 'Segmented neutrophils num (1000 cell/uL)',
    'LBDEONO': 'Eosinophils number (1000 cells/uL)',
    'LBDBANO': 'Basophils number (1000 cells/uL)',
    'LBXRBCSI': 'Red blood cell count (million cells/uL)',
    'LBXHGB': 'Hemoglobin (g/dL)',
    'LBXHCT': 'Hematocrit (%)',
    'LBXMCVSI': 'Mean cell volume (fL)',
    'LBXMCHSI': 'Mean cell hemoglobin (pg)',
    'LBXMC': 'Mean Cell Hgb Conc. (g/dL)',
    'LBXRDW': 'Red cell distribution width (%)',
    'LBXPLTSI': 'Platelet count (1000 cells/uL)',
    'LBXMPSI': 'Mean platelet volume (fL)'
}

df_CBC = df_CBC[columns_CBC]
df_CBC = df_CBC.rename(columns=rename_CBC)
df_CBC['SEQN'] = df_CBC['SEQN'].astype(int)
df_CBC = df_CBC.set_index('SEQN')
df_CBC.dropna(how='all', inplace=True)
df_CBC

check_coverage(df_demo, df_CBC)

df_medical = pd.merge(df_SBP, df_CBC, left_index=True, right_index=True, how='outer')

df_medical.to_csv('../processed_data/medical_table.csv')

### 4. Dietary Data

type_dietary = 'dietary'

#### 4.1. Total Nutrition Intakes

df_TOT1 = concat_data_across_years(type_dietary, 'DR1TOT', years, year_char)
df_TOT2 = concat_data_across_years(type_dietary, 'DR2TOT', years, year_char)

nutrition_facts = [
    "SEQN",
    "DBQ095Z", "DBD100", "DRQSPREP", "DR1STY", "DR1SKY", "DRQSDIET",
    "DRQSDT1", "DRQSDT2", "DRQSDT3", "DRQSDT4", "DRQSDT5", "DRQSDT6",
    "DRQSDT7", "DRQSDT8", "DRQSDT9", "DRQSDT10", "DRQSDT11", "DRQSDT12",
    "DRQSDT91", "DR1TNUMF", "DR1TWSZ", "DR1_300", "DRD340", "DRD360"
]
nutrition_data_1 = [
    "SEQN",
    "DR1TKCAL", "DR1TPROT", "DR1TCARB", "DR1TSUGR", "DR1TFIBE", "DR1TTFAT",
    "DR1TSFAT", "DR1TMFAT", "DR1TPFAT", "DR1TCHOL", "DR1TATOC", "DR1TATOA",
    "DR1TRET", "DR1TVARA", "DR1TACAR", "DR1TBCAR", "DR1TCRYP", "DR1TLYCO",
    "DR1TLZ", "DR1TVB1", "DR1TVB2", "DR1TNIAC", "DR1TVB6", "DR1TFOLA",
    "DR1TFA", "DR1TFF", "DR1TFDFE", "DR1TCHL", "DR1TVB12", "DR1TB12A",
    "DR1TVC", "DR1TVD", "DR1TVK", "DR1TCALC", "DR1TPHOS", "DR1TMAGN",
    "DR1TIRON", "DR1TZINC", "DR1TCOPP", "DR1TSODI", "DR1TPOTA", "DR1TSELE",
    "DR1TCAFF", "DR1TTHEO", "DR1TALCO", "DR1TMOIS", "DR1_320Z", "DR1_330Z",
    "DR1BWATZ"
]
nutrition_data_2 = [
    "SEQN",
    "DR2TKCAL", "DR2TPROT",
    "DR2TCARB", "DR2TSUGR", "DR2TFIBE", "DR2TTFAT", "DR2TSFAT", "DR2TMFAT",
    "DR2TPFAT", "DR2TCHOL", "DR2TATOC", "DR2TATOA", "DR2TRET", "DR2TVARA",
    "DR2TACAR", "DR2TBCAR", "DR2TCRYP", "DR2TLYCO", "DR2TLZ", "DR2TVB1",
    "DR2TVB2", "DR2TNIAC", "DR2TVB6", "DR2TFOLA", "DR2TFA", "DR2TFF",
    "DR2TFDFE", "DR2TCHL", "DR2TVB12", "DR2TB12A", "DR2TVC", "DR2TVD",
    "DR2TVK", "DR2TCALC", "DR2TPHOS", "DR2TMAGN", "DR2TIRON", "DR2TZINC",
    "DR2TCOPP", "DR2TSODI", "DR2TPOTA", "DR2TSELE", "DR2TCAFF", "DR2TTHEO",
    "DR2TALCO", "DR2TMOIS", "DR2_320Z", "DR2_330Z", "DR2BWATZ"
]

df_NF = df_TOT1[nutrition_facts]
df_ND1 = df_TOT1[nutrition_data_1]
df_ND2 = df_TOT2[nutrition_data_2]

df_ND = pd.DataFrame(np.vstack((df_ND1.to_numpy(), df_ND2.to_numpy())), columns=df_ND1.columns)
df_ND = df_ND.groupby('SEQN').mean().reset_index().fillna(0)

df_ND = df_ND.set_index('SEQN')
standard_scaler = StandardScaler()
minmax_scaler = MinMaxScaler()
df_standardized = pd.DataFrame(standard_scaler.fit_transform(df_ND), columns=df_ND.columns, index=df_ND.index)
df_ND = pd.DataFrame(minmax_scaler.fit_transform(df_standardized), columns=df_ND.columns, index=df_ND.index)

df_NF = df_NF.fillna(-1).astype(int)
df_NF = df_NF.set_index('SEQN')
df_N = pd.merge(df_ND, df_NF, left_index=True, right_index=True, how='outer')
df_N.index.name = 'SEQN'

df_N

df_N.to_csv('../processed_data/diet_table.csv')

check_coverage(df_demo, df_N)

### Miscellaneous Analysis - Broken below

df_IFF1 = concat_data_across_years(type_dietary, 'DR1IFF', years, year_char)
df_IFF2 = concat_data_across_years(type_dietary, 'DR2IFF', years, year_char)

# See here that not all our target has detailed food records. And this rate is 88.45%
df_temp = df_IFF1.set_index('SEQN')
check_coverage(df_demo, df_temp)

# df_IFF = pd.concat([df_IFF1, df_IFF2])
# df_IFF = df_IFF.groupby('SEQN').count().reset_index()
# temp = df_demo.merge(df_IFF, on='SEQN', how='left')
# temp = temp[['SEQN', 'label', 'Food code']]
# temp.groupby('label')['Food code'].mean()
