==> code/preprocessing/2.4_prompt_graph_construction.ipynb <==
from transformers import LlamaTokenizer, LlamaModel
import torch
from tqdm import tqdm
import numpy as np

tokenizer = LlamaTokenizer.from_pretrained('../llama-2-7b')
model = LlamaModel.from_pretrained('../llama-2-7b')

def get_llama_embedding(sentence):
    inputs = tokenizer(sentence, return_tensors="pt")
    with torch.no_grad():
        outputs = model(**inputs)
    # Extract the hidden states (last layer)
    last_hidden_state = outputs.last_hidden_state
    # Average the hidden states to get sentence embedding
    sentence_embedding = torch.mean(last_hidden_state, dim=1).squeeze().numpy()

    return sentence_embedding


def assign_prompt_embedding(graph, node_type):
    prompted_embedding = []
    for prompt in tqdm(graph[node_type]['prompt']):
        prompted_embedding.append(get_llama_embedding(prompt))
    graph[node_type].prompt_embedding = torch.tensor(np.array(prompted_embedding), dtype=torch.float32)

graph = torch.load('../processed_data/heterogeneous_graph_768_no_med_with_prompt_10_imbalanced.pt')

#### Be careful of running the following cells, it takes a long time to run.

assign_prompt_embedding(graph, 'user')

assign_prompt_embedding(graph, 'food')

assign_prompt_embedding(graph, 'ingredient')

assign_prompt_embedding(graph, 'category')

assign_prompt_embedding(graph, 'habit')

torch.save(graph, '../processed_data/heterogeneous_graph_768_no_med_with_prompt_10_imbalanced.pt')
